+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * new `emscope pack -a, --about-file` option
 * renamed `emscope grab -v, --voltage` option
+* renamed `emscope pack -s, --status` option
+* renamed `emscope pack --restore` option
 
 ## VERSION-25.0.1